@@ -24,7 +24,6 @@
     "@types/react-relay": "^0.9.8",
     "babel-cli": "^6.18.0",
     "babel-eslint": "^7.1.1",
-    "babel-loader": "^6.2.10",
     "babel-plugin-module-resolver": "^2.4.0",
     "babel-plugin-syntax-async-functions": "^6.13.0",
     "babel-plugin-transform-class-properties": "^6.19.0",
@@ -45,15 +44,15 @@
     "storyshots": "orta/storyshots#49bfe52",
     "ts-babel-node": "orta/ts-babel-node#tsx",
     "ts-jest": "^18.0.1",
-    "ts-loader": "^1.3.3",
     "ts-node": "^2.0.0",
     "tslint": "^4.3.1",
-    "webpack": "^1.14.0",
+    "webpack": "v2.2.0-rc.5",
     "webpack-dev-middleware": "^1.9.0",
     "webpack-hot-middleware": "^2.14.0"
   },
   "dependencies": {
     "aphrodite": "^1.1.0",
+    "awesome-typescript-loader": "v3.0.0-beta.13",
     "express": "^4.14.0",
     "isomorphic-fetch": "^2.2.1",
     "isomorphic-relay": "^0.7.3",
